Doubly-Circular file updated

+++ Data-Structures-Algorithms/Linked Lists/3. Circular Linked Lists/2. Doubly Circular Linked List/Doubly-Circular-Basics.cpp
@@ -6,7 +6,7 @@ class Node {
 
 public:
 
-	float data;
+	int data;
 	Node* prev, * next;
 
 	// Constructor
